refactor(connection-priority): polish API and add example controls
- Clarify return-value semantics in README and JSDoc: true means the
  request was successfully initiated; the priority change is applied
  asynchronously. false on iOS, error, or disconnected device.
- Rename the manager wrapper parameter to `priority` for ergonomics.
- Drop the redundant SecurityException catch on Android (the generic
  Exception catch already covers it).
- Add High / Balanced / Low Power connection-priority test buttons to
  the example app, and wire example entry to App.
