+++ example/package.json
@@ -14,7 +14,7 @@
     "expo-status-bar": "~55.0.4",
     "react": "19.2.0",
     "react-native": "0.83.2",
-    "react-native-ble-nitro": "file:../react-native-ble-nitro-1.11.0.tgz",
+    "react-native-ble-nitro": "file:../react-native-ble-nitro-1.13.0.tgz",
     "react-native-nitro-modules": "^0.35.0",
     "react-native-safe-area-context": "~5.6.2"
   },
